# src/dal_obscura/interfaces/http/routes/settings.py
"""Runtime and authentication settings routes.

Example:
    ```python
    app.include_router(router(deps))
    ```
"""

from __future__ import annotations

from typing import cast

from fastapi import APIRouter, Depends

from dal_obscura.control import identity_attributes, workspace_service
from dal_obscura.control.access import ControlPlaneActor
from dal_obscura.interfaces.http.routes.deps import ControlPlaneDeps
from dal_obscura.interfaces.http.routes.schemas import (
    AttributeMappingPreviewRequest,
    AttributeMappingPreviewResponse,
    AuthProviderResponse,
    AuthProviderRevisionResponse,
    AuthProvidersRequest,
    RuntimeSettingsRequest,
    RuntimeSettingsResponse,
)


def router(deps: ControlPlaneDeps) -> APIRouter:
    """Builds workspace settings routes.

    Example:
        ```python
        settings_router = router(deps)
        ```
    """

    api = APIRouter()

    @api.get(
        "/v1/settings/runtime",
        response_model=RuntimeSettingsResponse | None,
        dependencies=[Depends(deps.require_admin)],
    )
    def get_workspace_runtime_settings() -> RuntimeSettingsResponse | None:
        return cast(
            RuntimeSettingsResponse | None,
            deps.with_transaction(
                lambda service: workspace_service.get_workspace_runtime_settings(service.session)
            ),
        )

    @api.get(
        "/v1/settings/auth-providers",
        response_model=list[AuthProviderResponse],
        dependencies=[Depends(deps.require_admin)],
    )
    def list_workspace_auth_providers() -> list[AuthProviderResponse]:
        return cast(
            list[AuthProviderResponse],
            deps.with_transaction(
                lambda service: workspace_service.list_workspace_auth_providers(service.session)
            ),
        )

    @api.get(
        "/v1/settings/auth-providers/revision",
        response_model=AuthProviderRevisionResponse,
        dependencies=[Depends(deps.require_admin)],
    )
    def workspace_auth_provider_revision() -> AuthProviderRevisionResponse:
        return cast(
            AuthProviderRevisionResponse,
            deps.with_transaction(
                lambda service: {
                    "revision": workspace_service.workspace_auth_provider_revision(service.session)
                }
            ),
        )

    @api.put(
        "/v1/settings/runtime",
        response_model=RuntimeSettingsResponse,
        dependencies=[Depends(deps.require_admin)],
    )
    def upsert_workspace_runtime_settings(
        request: RuntimeSettingsRequest,
        actor: ControlPlaneActor = Depends(deps.require_admin),  # noqa: B008
    ) -> RuntimeSettingsResponse:
        result = deps.with_transaction(
            lambda service: workspace_service.upsert_workspace_runtime_settings(
                service.session,
                ttl=request.ticket_ttl_seconds,
                max_tickets=request.max_tickets,
                max_ticket_exchanges=request.max_ticket_exchanges,
                path_rules=request.path_rules,
                expected_revision=request.expected_revision,
                actor_principal="system" if actor is None else actor.identity_key(),
            )
        )
        return cast(
            RuntimeSettingsResponse,
            result
            or {
                "ticket_ttl_seconds": request.ticket_ttl_seconds,
                "max_tickets": request.max_tickets,
                "max_ticket_exchanges": request.max_ticket_exchanges,
                "path_rules": request.path_rules,
                "revision": 0,
            },
        )

    @api.put(
        "/v1/settings/auth-providers",
        response_model=list[AuthProviderResponse],
        dependencies=[Depends(deps.require_admin)],
    )
    def replace_workspace_auth_providers(
        request: AuthProvidersRequest,
        actor: ControlPlaneActor = Depends(deps.require_admin),  # noqa: B008
    ) -> list[AuthProviderResponse]:
        return cast(
            list[AuthProviderResponse],
            deps.with_transaction(
                lambda service: workspace_service.replace_workspace_auth_providers(
                    service.session,
                    request.providers,
                    expected_revision=request.expected_revision,
                    actor_principal="system" if actor is None else actor.identity_key(),
                )
            )
            or [],
        )

    @api.post(
        "/v1/settings/auth-providers/{ordinal}/attribute-preview",
        response_model=AttributeMappingPreviewResponse,
        dependencies=[Depends(deps.require_admin)],
    )
    def preview_attribute_mapping(
        ordinal: int, request: AttributeMappingPreviewRequest
    ) -> dict[str, object]:
        return deps.with_transaction(
            lambda service: identity_attributes.preview_attributes(
                service.session, ordinal, request.claims, request.provider_args
            )
        )

    return api